requirements.txt
pyaudio>=0.2.14
vosk>=0.3.45
pyautogui>=0.9.54
pyttsx3>=2.90
pywin32>=306
psutil>=5.9
keyboard>=0.13.5
cryptography>=42.0
rapidfuzz>=3.6
pycaw>=20240210
screen-brightness-control>=0.22
pystray>=0.19
Pillow>=10.0
plyer>=2.1
requests>=2.31
pyperclip>=1.8
google-genai>=0.2.0
edge-tts>=6.1.12

# note: tkinter and winsound are built-in natively in Python (Windows)
# note: wake word detection uses Vosk directly
